simplify DMA feature and move NVM+RAMSIM into cfg file
