+++ examples/posix/wh_posix_client/wh_posix_client_cfg.c
@@ -1,3 +1,7 @@
+/*
+ * wolfHSM Client POSIX Example
+ */
+
 #include "wh_posix_cfg.h"
 #include "wh_posix_client_cfg.h"
 
@@ -21,7 +25,8 @@ whTransportClientCb shmCb = POSIX_TRANSPORT_SHM_CLIENT_CB;
 whTransportClientCb tcpCb = PTT_CLIENT_CB;
 
 #ifdef WOLFSSL_STATIC_MEMORY
-whTransportClientCb dmaCb = POSIX_TRANSPORT_DMA_CLIENT_CB;
+whTransportClientCb dmaCb = POSIX_TRANSPORT_SHM_CLIENT_CB;
+whClientDmaConfig dmaConfig;
 
 const word32 sizeList[] = {
     WH_POSIX_STATIC_MEM_SIZE_1, WH_POSIX_STATIC_MEM_SIZE_2,
@@ -38,26 +43,65 @@ const word32 distList[] = {
     WH_POSIX_STATIC_MEM_DIST_9
 };
 
+
+int Client_ExampleSetupDmaMemory(void* ctx, void* conf)
+{
+    void* dma;
+    size_t dmaSz;
+    WOLFSSL_HEAP_HINT* hint = NULL;
+    int ret;
+    whClientContext* client = (whClientContext*)ctx;
+    whClientConfig* c_conf  = (whClientConfig*)conf;
+    posixTransportShmContext* shmCtx;
+
+    shmCtx = (posixTransportShmContext*)c_conf->comm->transport_context;
+    ret = posixTransportShm_GetDma(shmCtx, &dma, &dmaSz);
+    if (ret != 0) {
+        printf("Failed to get DMA\n");
+        return -1;
+    }
+
+    ret = wc_LoadStaticMemory_ex(&hint, WH_POSIX_STATIC_MEM_LIST_SIZE,
+        sizeList, distList, dma, dmaSz, 0, 0);
+    if (ret != 0) {
+        printf("Failed to load static memory\n");
+        return -1;
+    }
+    void* test = XMALLOC(1, hint, DYNAMIC_TYPE_TMP_BUFFER);
+    XFREE(test, hint, DYNAMIC_TYPE_TMP_BUFFER);
+
+    ret = wh_Client_SetHeap(client, (void*)hint);
+    if (ret != 0) {
+        printf("Failed to set heap\n");
+        return -1;
+    }
+
+    return 0;
+}
+
+
 /* client configuration setup example for transport */
 int Client_ExampleDMAConfig(void* conf)
 {
     whClientConfig* c_conf = (whClientConfig*)conf;
 
-    memset(&tccShm, 0, sizeof(posixTransportRefClientContext));
+    memset(&tccShm, 0, sizeof(posixTransportShmClientContext));
     memset(&c_comm, 0, sizeof(whCommClientConfig));
 
     shmConfig.name      = WH_POSIX_SHARED_MEMORY_NAME;
     shmConfig.req_size  = WH_POSIX_REQ_SIZE;
     shmConfig.resp_size = WH_POSIX_RESP_SIZE;
     shmConfig.dma_size  = WH_POSIX_DMA_SIZE;
-    shmConfig.dmaStaticMemListSz = WH_POSIX_STATIC_MEM_LIST_SIZE;
-    shmConfig.dmaStaticMemList = sizeList;
-    shmConfig.dmaStaticMemDist = distList;
+
+    dmaConfig.cb = wh_Client_PosixStaticMemoryDMA;
+    dmaConfig.dmaAddrAllowList = NULL;
 
     c_comm.transport_cb      = &dmaCb;
     c_comm.transport_context = (void*)&tccShm;
     c_comm.transport_config  = (void*)&shmConfig;
     c_comm.client_id         = WH_POSIX_CLIENT_ID;
+
+    c_conf->dmaConfig        = &dmaConfig;
     c_conf->comm             = &c_comm;
 
     return WH_ERROR_OK;
@@ -89,7 +133,7 @@ int Client_ExampleSHMConfig(void* conf)
 {
     whClientConfig* c_conf = (whClientConfig*)conf;
 
-    memset(&tccShm, 0, sizeof(posixTransportRefClientContext));
+    memset(&tccShm, 0, sizeof(posixTransportShmClientContext));
     memset(&c_comm, 0, sizeof(whCommClientConfig));
 
     shmConfig.name      = WH_POSIX_SHARED_MEMORY_NAME;
