+++ examples/posix/wh_posix_client/wh_posix_client_cfg.h
@@ -4,5 +4,5 @@
 int Client_ExampleDMAConfig(void* c_conf);
 int Client_ExampleSHMConfig(void* c_conf);
 int Client_ExampleTCPConfig(void* c_conf);
-
+int Client_ExampleSetupDmaMemory(void* ctx, void* c_conf);
 #endif /* WH_POSIX_CLIENT_CFG_H */
