+++ src/wh_server.c
@@ -103,12 +103,6 @@ int wh_Server_Init(whServerContext* server, whServerConfig* config)
     }
 
 #ifdef WOLFHSM_CFG_DMA
-    /* Check if this is a DMA transport and set up callbacks if needed */
-    if (wh_Server_IsDmaTransport(server->comm)) {
-        /* This is a DMA transport - set up DMA callback if not already set */
-        server->dma.cb = wh_Server_PosixStaticMemoryDMA;
-    }
-
     /* Initialize DMA configuration and callbacks, if provided */
     if (NULL != config->dmaConfig) {
         server->dma.dmaAddrAllowList = config->dmaConfig->dmaAddrAllowList;
@@ -132,22 +126,6 @@ int wh_Server_Cleanup(whServerContext* server)
     return WH_ERROR_OK;
 }
 
-
-int wh_Server_IsDmaTransport(whCommServer* comm)
-{
-    if (comm == NULL || comm->transport_context == NULL) {
-        return 0;
-    }
-
-#ifdef WOLFHSM_CFG_DMA
-    /* Check if this is a DMA transport by looking for heap hint */
-    void* heap_hint = NULL;
-    posixTransportShm_GetHeapHint(comm->transport_context, &heap_hint);
-    return (heap_hint != NULL) ? 1 : 0;
-#else
-    return 0;
-#endif /* WOLFHSM_CFG_DMA */
-}
 
 int wh_Server_SetConnected(whServerContext *server, whCommConnected connected)
 {
